requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
drf-spectacular>=0.27,<0.28